Fixes test output on Challenge 4
Decrypted plaintext ends with a newline character. Now instead of
printing the decrypted value, the test case prints the string
representation of it (i.e., "\n" instead of an actual line break).

Standardizing a bit as working on Challenge 6 led to spitting out a pile
of control characters that borked the terminal. Now I'm Doing The Right
Thing.